% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{attachmentBlock}
\alias{attachmentBlock}
\title{AdminLTE3 attachment container}
\usage{
attachmentBlock(..., image, title = NULL, href = NULL)
}
\arguments{
\item{...}{Any element.}

\item{image}{url or path to the image.}

\item{title}{Attachment title.}

\item{href}{External link.}
}
\description{
\link{attachmentBlock} create an attachment container, nice to wrap articles...
and insert in a \link{box}.
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)

 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "attachmentBlock example",
      attachmentBlock(
       image = "https://adminlte.io/themes/v3/dist/img/user1-128x128.jpg",
       title = "Test",
       href = "https://google.com",
       "This is the content"
      )
     )
    ),
    title = "attachmentBlock"
  ),
  server = function(input, output) { }
 )
}

}
\seealso{
Other boxWidgets: 
\code{\link{bs4CardLabel}()},
\code{\link{bs4CardSidebar}()},
\code{\link{bs4Carousel}()},
\code{\link{bs4SocialCard}()},
\code{\link{bs4Timeline}()},
\code{\link{cardDropdown}()},
\code{\link{cardProfile}()},
\code{\link{descriptionBlock}()},
\code{\link{userPost}()}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
\concept{boxWidgets}
