% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{dashboardBadge}
\alias{dashboardBadge}
\title{Alias to \link{bs4Badge}
See \link{bs4Badge} for full details}
\usage{
dashboardBadge(..., color, position = c("left", "right"), rounded = FALSE)
}
\description{
Alias to \link{bs4Badge}
See \link{bs4Badge} for full details
}
