% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{bs4Alert}
\alias{bs4Alert}
\title{Create a Bootstrap 4 alert}
\usage{
bs4Alert(
  ...,
  id = NULL,
  title,
  closable = TRUE,
  width = 6,
  elevation = NULL,
  status = c("primary", "warning", "danger", "info", "success")
)
}
\arguments{
\item{...}{Alert content.}

\item{id}{Alert id. Needed by \link{bs4CloseAlert}.}

\item{title}{Alert title.}

\item{closable}{Whether to allow the user to close the alert. FALSE by default.}

\item{width}{Alert width. Between 1 and 12.}

\item{elevation}{Alert elevation.}

\item{status}{Alert status. "primary", "success", "warning", "danger" or "info".}
}
\description{
AdminLTE3 alert
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     controlbar = dashboardControlbar(),
     footer = dashboardFooter(),
     title = "Alerts",
     body = dashboardBody(
       bs4Alert(
        title = "Be Careful!",
        status = "danger",
        closable = FALSE,
        "Danger alert preview. This alert is not dismissable. 
        A wonderful serenity has taken possession of my entire soul, 
        like these sweet mornings of spring which 
        I enjoy with my whole heart."
       ),
       bs4Alert(
        title = "Congratulation!",
        status = "success",
        closable = TRUE,
        elevation = 4,
        "Warning alert preview. This alert is dismissable. 
        A wonderful serenity has taken possession of my entire soul, 
        like these sweet mornings of spring which 
        I enjoy with my whole heart."
       )
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
