% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4Badge}
\alias{bs4Badge}
\title{Create a Bootstrap 4 dashboard badge item}
\usage{
bs4Badge(..., position = c("left", "right"), status, rounded = FALSE)
}
\arguments{
\item{...}{Badge content.}

\item{position}{Badge position: "left" or "right".}

\item{status}{Bdage color. "primary", "danger", "info", "success", "warning",
"secondary", "dark" or "light".}

\item{rounded}{Whether the badge is rounded instead of square. FALSE by default.}
}
\description{
Beautiful badge from AdminLTE3
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shiny::shinyApp(
   ui = bs4DashPage(
     navbar = bs4DashNavbar(),
     sidebar = bs4DashSidebar(),
     controlbar = bs4DashControlbar(),
     footer = bs4DashFooter(),
     title = "test",
     body = bs4DashBody(
      bs4Badge(
       position = "right",
       status = "warning",
       "Warning"
      )
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
