% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{box}
\alias{box}
\title{Alias to \link{bs4Card}
See \link{bs4Card} for full details}
\usage{
box(
  ...,
  title = NULL,
  footer = NULL,
  status = NULL,
  solidHeader = FALSE,
  background = NULL,
  width = 6,
  height = NULL,
  collapsible = TRUE,
  collapsed = FALSE,
  closable = FALSE,
  icon = NULL,
  gradient = FALSE,
  cardToolSize = "sm",
  maximizable = FALSE,
  elevation = NULL,
  headerBorder = TRUE,
  label = NULL,
  dropdownMenu = NULL,
  sidebar = NULL,
  id = NULL
)
}
\description{
Alias to \link{bs4Card}
See \link{bs4Card} for full details
}
