% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{cardSidebar}
\alias{cardSidebar}
\title{Alias to \link{bs4CardSidebar}
See \link{bs4CardSidebar} for full details}
\usage{
cardSidebar(
  ...,
  inputId = NULL,
  width = "25\%",
  background = "#333a40",
  startOpen = FALSE,
  icon = "cogs"
)
}
\description{
Alias to \link{bs4CardSidebar}
See \link{bs4CardSidebar} for full details
}
