% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4CarouselItem}
\alias{bs4CarouselItem}
\title{Create a Bootstrap 4 carousel item}
\usage{
bs4CarouselItem(active = FALSE, src = NULL)
}
\arguments{
\item{active}{Whether the item is active or not at start.}

\item{src}{Item path or url.}
}
\description{
To insert in a \link{bs4Carousel}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
