% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{bs4CloseAlert}
\alias{bs4CloseAlert}
\title{Close AdminLTE3 alert}
\usage{
bs4CloseAlert(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{\link{bs4Alert} id.}

\item{session}{Shiny session object.}
}
\description{
Server side function
}
\note{
One may use input$<id>, where id is the alert unique id, to trigger
more actions on the server side after the alert closed.
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shiny::shinyApp(
  ui = dashboardPage(
    navbar = dashboardHeader(),
    sidebar = dashboardSidebar(),
    body = dashboardBody(
      actionButton("close", "Close Alert"),
      bs4Alert(id = "myalert", title = "Hello", status = "success")
    ),
    controlbar = dashboardControlbar()
  ),
  server = function(input, output, session) {
    observeEvent(input$close, {
      bs4CloseAlert(id = "myalert")
    })
    
    observe(print(input$myalert))
    
    observeEvent(input$myalert, {
      bs4Toast(title = "Alert succesfully closed!")
    })
  }
 )
}
}
