% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4Dash-package.r
\docType{package}
\name{bs4Dash-package}
\alias{bs4Dash}
\alias{bs4Dash-package}
\title{bs4Dash: A 'Bootstrap 4' Version of 'shinydashboard'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Make 'Bootstrap 4' Shiny dashboards. Use the full power of 'AdminLTE3', a dashboard template built on top of 'Bootstrap 4' \url{https://github.com/ColorlibHQ/AdminLTE}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RinteRface/bs4Dash}
  \item \url{https://bs4dash.rinterface.com/}
  \item Report bugs at \url{https://github.com/RinteRface/bs4Dash/issues}
}

}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com}

Other contributors:
\itemize{
  \item  RinteRface [copyright holder]
  \item  Almasaeed Studio (AdminLTE3 theme for Bootstrap 4) [contributor, copyright holder]
  \item Winston Chang (Utils functions from shinydashboard) [contributor, copyright holder]
}

}
\keyword{internal}
