% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4DashAccordionItem}
\alias{bs4DashAccordionItem}
\title{Create a Bootstrap 4 accordion item}
\usage{
bs4DashAccordionItem(..., id, title = NULL, status = NULL, width = 6)
}
\arguments{
\item{...}{Item content.}

\item{id}{Item unique id.}

\item{title}{Item title.}

\item{status}{Item color. "primary", "success", "warning", "danger". NULL by default.}

\item{width}{The width of the accordion.}
}
\description{
To insert in a bs4DashAccordion
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
