% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4DashBadge}
\alias{bs4DashBadge}
\title{Create a Bootstrap 4 dashboard badge item}
\usage{
bs4DashBadge(..., position = c("left", "right"), status)
}
\arguments{
\item{...}{BDage content.}

\item{position}{Badge position: "left" or "right".}

\item{status}{Bdage color. "primary", "danger", "info", "success" or "warning"}
}
\description{
Beautiful badge from AdminLTE3
}
\examples{
if(interactive()){
 library(shiny)
 
 shiny::shinyApp(
   ui = bs4DashPage(
     navbar = bs4DashNavbar(),
     sidebar = bs4DashSidebar(),
     controlbar = bs4DashControlbar(),
     footer = bs4DashFooter(),
     title = "test",
     body = bs4DashBody(
      bs4DashBadge(
       position = "right",
       status = "warning",
       "Warning"
      )
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
