% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{bs4DashboardControlbar}
\alias{bs4DashboardControlbar}
\title{Create a Boostrap 4 dashboard right sidebar}
\usage{
bs4DashboardControlbar(..., skin = "dark", title = NULL)
}
\arguments{
\item{...}{Any UI element.}

\item{skin}{Controlbar skin. "dark" or "light".}

\item{title}{Controlbar title.}
}
\description{
Build an adminLTE3 dashboard right sidebar
}
\examples{

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
