% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{bs4DashboardFooter}
\alias{bs4DashboardFooter}
\title{Create a Boostrap 4 dashboard footer}
\usage{
bs4DashboardFooter(..., copyrights = NULL, right_text = NULL)
}
\arguments{
\item{...}{Left text.}

\item{copyrights}{Copyrights, if any.}

\item{right_text}{Right text.}
}
\description{
Build an adminLTE3 dashboard footer
}
\examples{

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
