% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardPage.R
\name{bs4DashboardPage}
\alias{bs4DashboardPage}
\title{Create a Boostrap 4 dashboard page}
\usage{
bs4DashboardPage(..., navbar = NULL, sidebar = NULL, controlbar = NULL,
  footer = NULL, title = NULL)
}
\arguments{
\item{...}{Body content.}

\item{navbar}{Bootstrap 4 dashboard navbar.}

\item{sidebar}{Bootstrap 4 dashboard main sidebar.}

\item{controlbar}{Bootstrap 4 dashboard control sidebar (right side).}

\item{footer}{Bootstrap 4 dashboard footer.}

\item{title}{App title.}
}
\description{
Build an adminLTE3 dashboard page
}
\examples{
if(interactive()){
 library(shiny)

 shiny::shinyApp(
   ui = bs4DashboardPage(
    navbar = bs4DashboardNavbar(),
    sidebar = bs4DashboardSidebar(),
    controlbar = bs4DashboardControlbar(),
    footer = bs4DashboardFooter(),
    title = "test"
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
