% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{bs4DashboardSidebar}
\alias{bs4DashboardSidebar}
\title{Create a Boostrap 4 dashboard main sidebar}
\usage{
bs4DashboardSidebar(..., title = NULL, skin = "dark", status = "primary",
  brandColor = NULL, url = NULL, src = NULL, elevation = 4,
  opacity = 0.8)
}
\arguments{
\item{...}{Slot for bs4SidebarMenu.}

\item{title}{Sidebar title.}

\item{skin}{Sidebar skin. "dark" or "light"}

\item{status}{Sidebar status. "primary", "danger", "warning", "success", "info".}

\item{brandColor}{Brand color. NULL by default: "primary", "danger", "warning",
"success", "info", "white" or "gray-light".}

\item{url}{Sidebar brand link.}

\item{src}{Sidebar brand image.}

\item{elevation}{Sidebar elevation. 4 by default.}

\item{opacity}{Sidebar opacity. From 0 to 1. 0.8 by default.}
}
\description{
Build an adminLTE3 dashboard main sidebar
}
\examples{

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
