% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4DashCarouselItem}
\alias{bs4DashCarouselItem}
\title{Create a Bootstrap 4 carousel item}
\usage{
bs4DashCarouselItem(active = FALSE, src = NULL)
}
\arguments{
\item{active}{Whether the item is active or not at start.}

\item{src}{Item path or url.}
}
\description{
To insert in a bs4DashCarousel
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
