% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R, R/aliases.R
\name{bs4DashControlbar}
\alias{bs4DashControlbar}
\alias{dashboardControlbar}
\title{Create a Boostrap 4 dashboard right sidebar}
\usage{
bs4DashControlbar(
  ...,
  id = NULL,
  disable = FALSE,
  width = 250,
  collapsed = TRUE,
  overlay = TRUE,
  skin = "dark",
  pinned = NULL
)

dashboardControlbar(
  ...,
  id = NULL,
  disable = FALSE,
  width = 250,
  collapsed = TRUE,
  overlay = TRUE,
  skin = "dark",
  pinned = NULL
)
}
\arguments{
\item{...}{Any UI element.}

\item{id}{To access the current state of the controlbar. Open is TRUE, closed
is FALSE. NULL by default.}

\item{disable}{If \code{TRUE}, the sidebar will be disabled.}

\item{width}{Controlbar width. 250 px by default.}

\item{collapsed}{Whether the control bar on the right side is collapsed or not at start. TRUE by default.}

\item{overlay}{Whether the sidebar covers the content when expanded. Default to TRUE.}

\item{skin}{Controlbar skin. "dark" or "light".}

\item{pinned}{Whether to block the controlbar state (TRUE or FALSE). Default to NULL.}
}
\description{
Build an adminLTE3 dashboard right sidebar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
