% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4DashGallery.R
\name{bs4DashGallery}
\alias{bs4DashGallery}
\title{Launch the bs4Dash Gallery}
\usage{
bs4DashGallery()
}
\description{
A gallery of all components available in bs4Dash.
}
\examples{

if (interactive()) {

 bs4DashGallery()

}
}
