% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R, R/aliases.R
\name{bs4DashNavbar}
\alias{bs4DashNavbar}
\alias{dashboardHeader}
\title{Boostrap 4 dashboard navbar}
\usage{
bs4DashNavbar(
  ...,
  title = NULL,
  titleWidth = NULL,
  disable = FALSE,
  .list = NULL,
  leftUi = NULL,
  rightUi = NULL,
  skin = "light",
  status = NULL,
  border = TRUE,
  compact = FALSE,
  sidebarIcon = shiny::icon("bars"),
  controlbarIcon = shiny::icon("th"),
  fixed = FALSE,
  fullscreen = FALSE
)

dashboardHeader(
  ...,
  title = NULL,
  titleWidth = NULL,
  disable = FALSE,
  .list = NULL,
  leftUi = NULL,
  rightUi = NULL,
  skin = "light",
  status = NULL,
  border = TRUE,
  compact = FALSE,
  sidebarIcon = shiny::icon("bars"),
  controlbarIcon = shiny::icon("th"),
  fixed = FALSE,
  fullscreen = FALSE
)
}
\arguments{
\item{...}{Any UI element between left and right Ui.}

\item{title}{Dashboard title (displayed top-left side). Alternatively, use \link{bs4DashBrand}
for more evolved title.}

\item{titleWidth}{The width of the title area. This must either be a number
which specifies the width in pixels, or a string that specifies the width
in CSS units.}

\item{disable}{If \code{TRUE}, don't display the header bar.}

\item{.list}{An optional list containing items to put in the header. Same as
the \code{...} arguments, but in list format. This can be useful when
working with programmatically generated items.}

\item{leftUi}{Custom left Ui content. Any element like \link{dropdownMenu}.}

\item{rightUi}{Custom right Ui content. Any element like \link{dropdownMenu}.}

\item{skin}{Navbar skin. "dark" or "light".}

\item{status}{Navbar status. "primary", "danger", "warning",
"success", "info", "white", "gray-light" and all other available colors. See
\link{getAdminLTEColors}.}

\item{border}{Whether to separate the navbar and body by a border. TRUE by default.}

\item{compact}{Whether items should be compacted. FALSE by default.}

\item{sidebarIcon}{Icon of the main sidebar toggle.}

\item{controlbarIcon}{Icon to toggle the controlbar (left).}

\item{fixed}{Whether to fix the navbar to the top. FALSE by default.}

\item{fullscreen}{Whether to allow fullscreen feature in the navbar. Default to FALSE.}
}
\description{
\link{dashboardHeader} creates an adminLTE3 dashboard navbar to be included in
\link{dashboardPage}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
