% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R, R/aliases.R
\name{bs4DashSidebar}
\alias{bs4DashSidebar}
\alias{dashboardSidebar}
\title{Create a Boostrap 4 dashboard main sidebar}
\usage{
bs4DashSidebar(
  ...,
  disable = FALSE,
  width = NULL,
  skin = "dark",
  status = "primary",
  elevation = 4,
  collapsed = FALSE,
  minified = TRUE,
  expandOnHover = TRUE,
  fixed = TRUE,
  id = NULL
)

dashboardSidebar(
  ...,
  disable = FALSE,
  width = NULL,
  skin = "dark",
  status = "primary",
  elevation = 4,
  collapsed = FALSE,
  minified = TRUE,
  expandOnHover = TRUE,
  fixed = TRUE,
  id = NULL
)
}
\arguments{
\item{...}{Slot for \link{bs4SidebarMenu}.}

\item{disable}{If \code{TRUE}, the sidebar will be disabled.}

\item{width}{The width of the sidebar. This must either be a number which
specifies the width in pixels, or a string that specifies the width in CSS
units.}

\item{skin}{Sidebar skin. "dark" or "light".}

\item{status}{Sidebar status. "primary", "danger", "warning", "success", "info".}

\item{elevation}{Sidebar elevation. 4 by default (until 5).}

\item{collapsed}{If \code{TRUE}, the sidebar will be collapsed on app startup.}

\item{minified}{Whether to slightly close the sidebar but still show item icons. Default
to TRUE.}

\item{expandOnHover}{Whether to expand the sidebar om hover. TRUE by default.}

\item{fixed}{Whether to fix the sidebar. Default to TRUE.}

\item{id}{Recover the state of the sidebar.}
}
\description{
Build an adminLTE3 dashboard main sidebar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
