% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{infoBoxOutput}
\alias{infoBoxOutput}
\title{Alias to \link{bs4InfoBoxOutput}
See \link{bs4InfoBoxOutput} for full details}
\usage{
infoBoxOutput(outputId, width = 4)
}
\arguments{
\item{outputId}{Output variable name.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}
}
\description{
Alias to \link{bs4InfoBoxOutput}
See \link{bs4InfoBoxOutput} for full details
}
