% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4ListGroupItem}
\alias{bs4ListGroupItem}
\title{BS4 list group item for AdminLTE3}
\usage{
bs4ListGroupItem(
  ...,
  active = FALSE,
  disabled = FALSE,
  type = c("basic", "action", "heading"),
  src = "#",
  title = NULL,
  subtitle = NULL,
  footer = NULL
)
}
\arguments{
\item{...}{Item content.}

\item{active}{Whether the item is active or not. FALSE by default. 
Only if type is "action" or "heading".}

\item{disabled}{Whether the item is disabled or not. FALSE by default. 
Only if type is "action" or "heading".}

\item{type}{Item type. Choose between "basic", "action" and "heading".}

\item{src}{Item external link.}

\item{title}{Item title (only if type is "heading").}

\item{subtitle}{Item subtitle (only if type is "heading").}

\item{footer}{Item footer content (only if type is "heading").}
}
\description{
Create a list group item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
