% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4Loading}
\alias{bs4Loading}
\title{AdminLTE3 loading state element}
\usage{
bs4Loading()
}
\description{
When a section is still work in progress or a computation is running
}
\note{
Loading state can be programmatically used when a conputation is running for instance.
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shiny::shinyApp(
   ui = bs4DashPage(
     navbar = bs4DashNavbar(),
     sidebar = bs4DashSidebar(),
     controlbar = bs4DashControlbar(),
     footer = bs4DashFooter(),
     title = "test",
     body = bs4DashBody(
       title = "Callouts",
       bs4Card(
        title = "Loading State",
        bs4Loading()
       )
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
