% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{bs4PopoverServer}
\alias{bs4PopoverServer}
\title{Create a Bootstrap 4 Popover from the server side}
\usage{
bs4PopoverServer(
  id = NULL,
  selector = NULL,
  options,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Popover target id.}

\item{selector}{jQuery selector. Allow more customization for the target (nested tags).}

\item{options}{List of options to pass to the popover. See \url{https://getbootstrap.com/docs/4.0/components/popovers/}.}

\item{session}{Shiny session object.}
}
\description{
This replaces the shinyBS popover feature that is not compatible
with Bootstrap 4
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(),
    sidebar = dashboardSidebar(),
    controlbar = dashboardControlbar(),
    footer = dashboardFooter(),
    title = "Popover server",
    body = dashboardBody(
     actionButton("goButton", "Show popover!"),
     actionButton("goButton2", "You can't see me first!")
    )
  ),
  server = function(input, output, session) {
   observeEvent(input$goButton, {
      bs4PopoverServer(
        id = "goButton2", 
        options = list(
         content = "Vivamus sagittis lacus vel augue laoreet rutrum faucibus.",
         title = "Server popover", 
         placement = "bottom",
         trigger = "hover"
        )
      )
    })
  }
 )
}
}
