% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{bs4PopoverUI}
\alias{bs4PopoverUI}
\title{Create a Bootstrap 4 Popover from the UI side}
\usage{
bs4PopoverUI(
  tag,
  content,
  title,
  placement = c("top", "bottom", "left", "right")
)
}
\arguments{
\item{tag}{Popover target.}

\item{content}{Popover content.}

\item{title}{Popover title.}

\item{placement}{Popover placement: "top", "bottom", "left" or "right".}
}
\description{
This replaces the shinyBS popover feature that is not compatible
with Bootstrap 4
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shiny::shinyApp(
  ui = bs4DashPage(
    enable_preloader = TRUE,
    navbar = bs4DashNavbar(),
    sidebar = bs4DashSidebar(),
    controlbar = bs4DashControlbar(),
    footer = bs4DashFooter(),
    title = "test",
    body = bs4DashBody(
     bs4PopoverUI(
      actionButton("goButton", "Click me to see the popover!"),
       title = "My popover",
       placement = "right",
       content = "Vivamus sagittis lacus vel augue laoreet rutrum faucibus."
     )
    )
  ),
  server = function(input, output) {}
 )
}
}
