% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4ProgressBar}
\alias{bs4ProgressBar}
\title{AdminLTE3 progress bar}
\usage{
bs4ProgressBar(
  value,
  min = 0,
  max = 100,
  vertical = FALSE,
  striped = FALSE,
  animated = FALSE,
  status = "primary",
  size = NULL,
  label = NULL
)
}
\arguments{
\item{value}{Progress bar value.}

\item{min}{Progress bar minimum value.}

\item{max}{Progress bar maximum value.}

\item{vertical}{Whether to display the progress bar in vertical mode. FALSE by default.}

\item{striped}{Whether the progress bar is striped or not. FALSE by default.}

\item{animated}{Whether to animate the progress bar. Default to FALSE.}

\item{status}{Progress bar status. "primary", "success", "warning", "danger" or "info".}

\item{size}{Progress bar size. NULL, "sm", "xs" or "xxs".}

\item{label}{Progress label. NULL by default.}
}
\description{
Create a Bootstrap 4 progress bar.
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     body = dashboardBody(
      box(
       title = "Horizontal",
       progressBar(
        value = 10,
        striped = TRUE,
        animated = TRUE
       ),
       progressBar(
        value = 50,
        status = "warning",
        size = "xs"
       ),
       progressBar(
        value = 20,
        status = "danger",
        size = "sm"
       )
      ),
      box(
       title = "Vertical",
       progressBar(
        value = 10,
        striped = TRUE,
        animated = TRUE,
        vertical = TRUE
       ),
       progressBar(
        value = 50,
        status = "warning",
        size = "xs",
        vertical = TRUE
       ),
       progressBar(
        value = 20,
        status = "danger",
        size = "sm",
        vertical = TRUE
       )
      )
     ),
     title = "Progress bars"
   ),
   server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
