% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R, R/aliases.R
\name{bs4SidebarHeader}
\alias{bs4SidebarHeader}
\alias{sidebarHeader}
\title{Create a Boostrap 4 dashboard main sidebar header}
\usage{
bs4SidebarHeader(title)

sidebarHeader(title)
}
\arguments{
\item{title}{SidebarHeader title.}
}
\description{
Build an adminLTE3 dashboard main sidebar header
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
