% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{bs4SidebarMenuItem}
\alias{bs4SidebarMenuItem}
\title{Create a Boostrap 4 dashboard main sidebar menu item}
\usage{
bs4SidebarMenuItem(..., tabName = NULL, icon = NULL, active = FALSE)
}
\arguments{
\item{...}{Item name.}

\item{tabName}{Should correspond exactly to the tabName given in \code{\link{bs4TabItem}}.}

\item{icon}{Item icon.}

\item{active}{Whether the item is active. FALSE by default.}
}
\description{
Build an adminLTE3 dashboard main sidebar menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
