% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{bs4SidebarMenuItemList}
\alias{bs4SidebarMenuItemList}
\title{Create a Boostrap 4 dashboard main sidebar menu item list}
\usage{
bs4SidebarMenuItemList(..., name = NULL, icon = NULL, open = FALSE,
  active = FALSE)
}
\arguments{
\item{...}{Slot for bs4SidebarMenuItem.}

\item{name}{Item list name.}

\item{icon}{Item list icon.}

\item{open}{Whether to display the item list in an open state. FALSE by default.}

\item{active}{Whether the section is active. FALSE by default.}
}
\description{
Build an adminLTE3 dashboard main sidebar menu item list
}
\examples{

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
