% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R, R/aliases.R
\name{bs4SidebarUserPanel}
\alias{bs4SidebarUserPanel}
\alias{sidebarUserPanel}
\title{Create a Boostrap 4 dashboard main sidebar user panel}
\usage{
bs4SidebarUserPanel(name, image = NULL)

sidebarUserPanel(name, image = NULL)
}
\arguments{
\item{name}{Name of the user.}

\item{image}{A filename or URL to use for an image of the person. If it is a
local file, the image should be contained under the www/ subdirectory of
the application.}
}
\description{
Build an adminLTE3 dashboard main sidebar user panel
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
