% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4Sortable}
\alias{bs4Sortable}
\title{BS4 sortable section}
\usage{
bs4Sortable(..., width = NULL)
}
\arguments{
\item{...}{Slot for UI elements such as \link{bs4Card}.}

\item{width}{Section width: between 1 and 12.}
}
\description{
Create a sortable UI section
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 shiny::shinyApp(
  ui = bs4DashPage(
    navbar = bs4DashNavbar(),
    sidebar = bs4DashSidebar(),
    controlbar = bs4DashControlbar(),
    footer = bs4DashFooter(),
    title = "test",
    body = bs4DashBody(
      bs4TabItems(
        bs4TabItem(
          tabName = "sortabled",
          fluidRow(
            lapply(1:3, FUN = function(i) {
              bs4Sortable(
                width = 4,
                p(class = "text-center", paste("Column", i)),
                lapply(1:2, FUN = function(j) {
                  bs4Card(
                    title = paste0("I am the ", j,"-th card of the ", i, "-th column"), 
                    width = 12,
                    "Click on my header"
                  )
                })
              )
            })
          )
        )
      )
    )
  ),
  server = function(input, output) {}
 )
}  

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
