% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4TableItem}
\alias{bs4TableItem}
\title{Create a Boostrap 4 table item}
\usage{
bs4TableItem(..., dataCell = FALSE)
}
\arguments{
\item{...}{Any HTML element.}

\item{dataCell}{Whether the cell should be contain data or text. <td> by default.}
}
\description{
Build an bs4 table item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
