% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4TableItems}
\alias{bs4TableItems}
\title{Create a Boostrap 4 table item row}
\usage{
bs4TableItems(...)
}
\arguments{
\item{...}{Slot for \link{bs4TableItem}.}
}
\description{
Build an bs4 table item row
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
