% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{bs4TabPanel}
\alias{bs4TabPanel}
\title{Create a tabPanel}
\usage{
bs4TabPanel(..., title, active = FALSE)
}
\arguments{
\item{...}{Tab content}

\item{title}{Tab title: it will be also passed as the id argument. Should be unique.}

\item{active}{Whether the tab is active or not. FALSE bu default.}
}
\description{
To be included in a bs4TabCard
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
