% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{bs4TabSetPanel}
\alias{bs4TabSetPanel}
\title{Create a tabSetPanel}
\usage{
bs4TabSetPanel(...)
}
\arguments{
\item{...}{Slot for bs4TabPanel.}
}
\description{
Imported by bs4TabCard. Do not use outside!
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
