% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4TimelineItemMedia}
\alias{bs4TimelineItemMedia}
\title{AdminLTE2 timeline media item}
\usage{
bs4TimelineItemMedia(src = NULL, height = NULL, width = NULL)
}
\arguments{
\item{src}{media url or path.}

\item{height}{media height in pixels.}

\item{width}{media width in pixels.}
}
\description{
Create a timeline media item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
