% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4TimelineLabel}
\alias{bs4TimelineLabel}
\title{AdminLTE3 timeline label}
\usage{
bs4TimelineLabel(..., status = NULL)
}
\arguments{
\item{...}{Any element.}

\item{status}{Label status: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create a timeline label
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
