% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4TimelineStart}
\alias{bs4TimelineStart}
\title{AdminLTE3 timeline starting point}
\usage{
bs4TimelineStart(icon = "clock-o", status = NULL)
}
\arguments{
\item{icon}{Item icon such as "clock-o", "times", ...}

\item{status}{Item status: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create a timeline starting point
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
