% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R, R/aliases.R
\name{bs4TooltipUI}
\alias{bs4TooltipUI}
\alias{tooltip}
\title{Create a Bootstrap 4 Tooltip from the UI side}
\usage{
bs4TooltipUI(tag, title, placement = c("top", "bottom", "left", "right"))

tooltip(tag, title, placement = c("top", "bottom", "left", "right"))
}
\arguments{
\item{tag}{Tooltip target.}

\item{title}{Tooltip title.}

\item{placement}{Tooltip placement: "top", "bottom", "left" or "right".}
}
\description{
This replaces the shinyBS tooltip feature that is not compatible
with Bootstrap 4
}
\note{
\link{tooltip} does not automatically handles tooltip removal and must be seperately implemented.
If the \link{dashboardHeader} help parameter is TRUE, all tooltips may be enabled
or disabled depending on the switch value, which may solve this problem. 
This allows to toggle tooltips whenever required.
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(),
    sidebar = dashboardSidebar(),
    controlbar = dashboardControlbar(),
    footer = dashboardFooter(),
    title = "Tooltip UI",
    body = dashboardBody(
     tooltip(
      actionButton("goButton", "Hover to see the tooltip"),
      title = "My tooltip",
      placement = "top"
     )
    )
  ),
  server = function(input, output) {}
 )
}
}
