% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{bs4TooltipServer}
\alias{bs4TooltipServer}
\title{Create a Bootstrap 4 Tooltip from the server side}
\usage{
bs4TooltipServer(
  id = NULL,
  selector = NULL,
  options,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Tooltip target id.}

\item{selector}{jQuery selector. Allow more customization for the target (nested tags).}

\item{options}{List of options to pass to the tooltip. See \url{https://getbootstrap.com/docs/4.0/components/tooltips/}.}

\item{session}{Shiny session object.}
}
\description{
This replaces the shinyBS tooltip feature that is not compatible
with Bootstrap 4
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(),
    sidebar = dashboardSidebar(),
    controlbar = dashboardControlbar(),
    footer = dashboardFooter(),
    title = "Tooltip server",
    body = bs4DashBody(
     actionButton("goButton", "Click on me to add tooltip"),
     actionButton("goButton2", "You can't see me first!")
    )
  ),
  server = function(input, output, session) {
   observeEvent(input$goButton, {
     bs4TooltipServer(
      id = "goButton2", 
      options = list(
       title = "Server tooltip", 
       placement = "bottom"
      )
     )
   })
  }
 )
}
}
