% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{bs4TooltipUI}
\alias{bs4TooltipUI}
\title{Create a Bootstrap 4 Tooltip from the UI side}
\usage{
bs4TooltipUI(tag, title, placement = c("top", "bottom", "left", "right"))
}
\arguments{
\item{tag}{Tooltip target.}

\item{title}{Tooltip title.}

\item{placement}{Tooltip placement: "top", "bottom", "left" or "right".}
}
\description{
This replaces the shinyBS tooltip feature that is not compatible
with Bootstrap 4
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = dashboardPage(
    enable_preloader = TRUE,
    header = dashboardHeader(),
    sidebar = dashboardSidebar(),
    controlbar = dashboardControlbar(),
    footer = dashboardFooter(),
    title = "Tooltip UI",
    body = dashboardBody(
     bs4TooltipUI(
      actionButton("goButton", "Hover to see the tooltip"),
      title = "My tooltip",
      placement = "top"
     )
    )
  ),
  server = function(input, output) {}
 )
}
}
