% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R, R/aliases.R
\name{bs4UserMenu}
\alias{bs4UserMenu}
\alias{dashboardUser}
\title{Create a Bootstrap 4 user profile.}
\usage{
bs4UserMenu(
  ...,
  name = NULL,
  image = NULL,
  title = NULL,
  subtitle = NULL,
  footer = NULL,
  status = NULL
)

dashboardUser(
  ...,
  name = NULL,
  image = NULL,
  title = NULL,
  subtitle = NULL,
  footer = NULL,
  status = NULL
)
}
\arguments{
\item{...}{Body content. Slot for \link{dashboardUserItem}.}

\item{name}{User name.}

\item{image}{User profile picture.}

\item{title}{A title.}

\item{subtitle}{A subtitle.}

\item{footer}{Footer is any.}

\item{status}{Ribbon status: "primary", "danger", "success", "warning", "info" and 
"secondary".}
}
\description{
Insert in the rightUi or leftUi slot of \link{dashboardHeader}.
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(rightUi = userOutput("user")),
    sidebar = dashboardSidebar(),
    body = dashboardBody(),
    title = "DashboardPage"
  ),
  server = function(input, output) {
   output$user <- renderUser({
    dashboardUser(
       name = "Divad Nojnarg", 
       image = "https://adminlte.io/themes/AdminLTE/dist/img/user2-160x160.jpg", 
       title = "shinydashboardPlus",
       subtitle = "Author", 
       footer = p("The footer", class = "text-center"),
       fluidRow(
        dashboardUserItem(
         width = 6,
         "Item 1"
        ),
        dashboardUserItem(
         width = 6,
         "Item 2"
        )
       )
      )
   })
  }
 )
}

}
