% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/aliases.R
\name{bs4ValueBox}
\alias{bs4ValueBox}
\alias{valueBox}
\title{Boostrap 4 value box}
\usage{
bs4ValueBox(
  value,
  subtitle,
  icon = NULL,
  color = NULL,
  width = 3,
  href = NULL,
  footer = NULL,
  elevation = NULL
)

valueBox(
  value,
  subtitle,
  icon = NULL,
  color = NULL,
  width = 3,
  href = NULL,
  footer = NULL,
  elevation = NULL
)
}
\arguments{
\item{value}{The value to display in the box. Usually a number or short text.}

\item{subtitle}{Subtitle text.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{href}{An optional URL to link to in the footer. Should both `footer`
and this parameter be set, `footer` will take precedence.}

\item{footer}{Optional html content for the footer of the box.}

\item{elevation}{Value box elevation.}

\item{status}{A color for the box. "primary", "info", "success", "warning", "danger" or NULL.}
}
\description{
A beautiful AdminLTE3 value box.
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     controlbar = dashboardControlbar(),
     footer = dashboardFooter(),
     title = "test",
     body = bs4DashBody(
      fluidRow(
       valueBox(
        value = 150,
        subtitle = "New orders",
        color = "primary",
        icon = icon("shopping-cart")
       ),
       valueBox(
        value = "53\%",
        subtitle = "New orders",
        color = "indigo",
        icon = icon("cogs"),
        footer = div("Hello World")
       ),
       valueBox(
        value = "44",
        subtitle = "User Registrations",
        color = "teal",
        icon = icon("sliders")
       )
      )
     )
   ),
   server = function(input, output) {}
 )
}

}
\seealso{
Other cards: 
\code{\link{bs4InfoBox}()},
\code{\link{bs4TabCard}()}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
\concept{cards}
