% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-functions.R, R/aliases.R
\name{bs4ValueBoxOutput}
\alias{bs4ValueBoxOutput}
\alias{bs4InfoBoxOutput}
\alias{valueBoxOutput}
\title{Create a value box output (client side)}
\usage{
bs4ValueBoxOutput(outputId, width = 4)

bs4InfoBoxOutput(outputId, width = 4)

valueBoxOutput(outputId, width = 4)
}
\arguments{
\item{outputId}{Output variable name.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}
}
\description{
This is the UI-side function for creating a dynamic \code{\link{bs4ValueBox}}.
}
\seealso{
\code{\link{renderbs4ValueBox}} for the corresponding server-side
  function and examples.
}
