% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cardComment}
\alias{cardComment}
\title{BS4 card comment container}
\usage{
cardComment(..., src = NULL, title = NULL, date = NULL)
}
\arguments{
\item{...}{Comment content.}

\item{src}{Author image, if any.}

\item{title}{Comment title.}

\item{date}{Date of publication.}
}
\description{
Create a card comment to insert in \link{bs4SocialCard}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
