% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/aliases.R
\name{cardDropdown}
\alias{cardDropdown}
\alias{boxDropdown}
\title{Create a box dropdown item list}
\usage{
cardDropdown(..., icon = shiny::icon("wrench"))

boxDropdown(..., icon = shiny::icon("wrench"))
}
\arguments{
\item{...}{Slot for \link{dropdownItem}.}

\item{icon}{Dropdown menu icon.}
}
\description{
Can be used to add dropdown items to a cardtool.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
