% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{boxDropdownItem}
\alias{boxDropdownItem}
\title{Alias to \link{cardDropdownItem}
See \link{cardDropdownItem} for full details}
\usage{
boxDropdownItem(..., id = NULL, href = NULL, icon = NULL)
}
\description{
Alias to \link{cardDropdownItem}
See \link{cardDropdownItem} for full details
}
