% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{boxLabel}
\alias{boxLabel}
\title{Alias to \link{cardLabel}
See \link{cardLabel} for full details}
\usage{
boxLabel(text, status, tooltip = NULL)
}
\description{
Alias to \link{cardLabel}
See \link{cardLabel} for full details
}
