% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cardProfile}
\alias{cardProfile}
\title{AdminLTE3 card profile}
\usage{
cardProfile(..., image = NULL, title, subtitle = NULL, bordered = FALSE)
}
\arguments{
\item{...}{Any element such as \link{cardProfileItemList}.}

\item{title}{Title.}

\item{subtitle}{Subtitle.}

\item{src}{Profile image, if any.}
}
\description{
Create card profile
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(
      sidebarMenu(
       sidebarHeader("Main content"),
       menuItem(
         "Profile Card",
         tabName = "profile_card",
         icon = icon("desktop")
       )
      )
     ),
     controlbar = dashboardControlbar(),
     footer = dashboardFooter(),
     title = "boxProfile",
     body = dashboardBody(
      tabItems(
       tabItem(
         tabName = "profile_card",
         bs4Card(
          status = "primary",
          solidHeader = TRUE,
          boxProfile(
           image = "https://adminlte.io/themes/AdminLTE/dist/img/user4-128x128.jpg",
           title = "Nina Mcintire",
           subtitle = "Software Engineer",
           bordered = TRUE,
           boxProfileItem(
             title = "Followers",
             description = 1322
            ),
            boxProfileItem(
             title = "Following",
             description = 543
            ),
            boxProfileItem(
             title = "Friends",
             description = 13287
            )
          )
        )
       )
      )
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
