% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cardProfileItem}
\alias{cardProfileItem}
\title{AdminLTE3 card profile item}
\usage{
cardProfileItem(title = NULL, description = NULL)
}
\arguments{
\item{title}{Item title.}

\item{description}{Item info.}
}
\description{
Create card profile item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
