% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cardProfileItemList}
\alias{cardProfileItemList}
\title{AdminLTE3 card profile item container}
\usage{
cardProfileItemList(..., bordered = FALSE)
}
\arguments{
\item{...}{Slot for \link{cardProfileItem}.}

\item{bordered}{Whether the container should have a border or not. FALSE by default.}
}
\description{
Create card profile item list
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
