% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col2css}
\alias{col2css}
\title{Create a CSS rgba declaration for a color name}
\usage{
col2css(color, alpha = NULL)
}
\arguments{
\item{color}{\code{chr} Color name, see \link[grDevices]{color}}

\item{alpha}{\code{num} alpha value}
}
\value{
\code{chr} css `rgba()` formatted color
}
\description{
Create a CSS rgba declaration for a color name
}
