% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R, R/aliases.R
\name{bs4DashBody}
\alias{bs4DashBody}
\alias{bs4TabItems}
\alias{bs4TabItem}
\alias{dashboardBody}
\alias{tabItems}
\alias{tabItem}
\title{Boostrap 4 dashboard body}
\usage{
bs4DashBody(...)

bs4TabItems(..., .list = NULL)

bs4TabItem(tabName = NULL, ...)

dashboardBody(...)

tabItems(..., .list = NULL)

tabItem(tabName = NULL, ...)
}
\arguments{
\item{...}{Contents of the tab.}

\item{.list}{Pass items as list with \link{lapply} family functions.}

\item{tabName}{The name of a tab. This must correspond to the \code{tabName}
of a sidebar \code{\link{menuItem}}.}
}
\description{
\link{dashboardBody} creates the main body container for a \link{dashboardPage}.

\link{tabItems} creates a wrapper for multiple \link{tabItem}.

\link{tabItem} creates a body tab content.
}
\seealso{
\link{dashboardSidebar}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
