% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R, R/aliases.R
\name{bs4DashFooter}
\alias{bs4DashFooter}
\alias{dashboardFooter}
\title{Dashboard Footer}
\usage{
bs4DashFooter(left = NULL, right = NULL, fixed = FALSE)

dashboardFooter(left = NULL, right = NULL, fixed = FALSE)
}
\arguments{
\item{left}{Left text.}

\item{right}{Right text.}

\item{fixed}{Whether to fix the footer. Default to FALSE.}
}
\description{
This creates a dashboard footer to insert in \link{dashboardPage}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
