% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{dropdownDivider}
\alias{dropdownDivider}
\title{Create a box dropdown divider}
\usage{
dropdownDivider()
}
\description{
Create a box dropdown divider
}
\note{
Useful to separate 2 sections of dropdown items.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
