% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{dropdownHeader}
\alias{dropdownHeader}
\title{Dropdown header helper}
\usage{
dropdownHeader(text)
}
\arguments{
\item{text}{Text to display.}
}
\value{
A shiny tag.
}
\description{
Display header text within dropdown menu
}
