% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{dropdownItem}
\alias{dropdownItem}
\title{Create a box dropdown item}
\usage{
dropdownItem(url = NULL, name = NULL)
}
\arguments{
\item{url}{Target url or page.}

\item{name}{Item name.}
}
\description{
Create a box dropdown item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
