% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R, R/aliases.R
\name{bs4Jumbotron}
\alias{bs4Jumbotron}
\alias{jumbotron}
\title{BS4 jumbotron for AdminLTE3}
\usage{
bs4Jumbotron(
  ...,
  title = NULL,
  lead = NULL,
  href = NULL,
  btnName = "More",
  status = c("primary", "warning", "danger", "info", "success")
)

jumbotron(
  ...,
  title = NULL,
  lead = NULL,
  href = NULL,
  btnName = "More",
  status = c("primary", "warning", "danger", "info", "success")
)
}
\arguments{
\item{...}{Any content.}

\item{title}{Jumbotron title.}

\item{lead}{Jumbotron lead.}

\item{href}{Jumbrotron external link.}

\item{btnName}{Jumbotron button name.}

\item{status}{Jumbotron background color. "primary", "success", "warning", "danger" or "info".}
}
\description{
Create a jumbotron
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     controlbar = dashboardControlbar(),
     footer = dashboardFooter(),
     title = "Jumbotron",
     body = dashboardBody(
      jumbotron(
      title = "Hello, world!",
      lead = "This is a simple hero unit, a simple jumbotron-style 
      component for calling extra attention to featured 
      content or information.",
      "It uses utility classes for typography and spacing 
      to space content out within the larger container.",
      status = "primary",
      href = "http://www.google.fr"
      )
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
