% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{messageItem}
\alias{messageItem}
\title{Bootstrap 4 message item}
\usage{
messageItem(
  from,
  message,
  icon = shiny::icon("user"),
  time = NULL,
  href = NULL,
  image = NULL,
  color = "secondary",
  inputId = NULL
)
}
\arguments{
\item{from}{Who the message is from.}

\item{message}{Text of the message.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{time}{String representing the time the message was sent. Any string may
be used. For example, it could be a relative date/time like "5 minutes",
"today", or "12:30pm yesterday", or an absolute time, like "2014-12-01 13:45".
If NULL, no time will be displayed.}

\item{href}{An optional URL to link to.}

\item{image}{User image.}

\item{color}{Item color.
If not NULL, Indicate the message priority.}

\item{inputId}{Whether to allow the item to act as a \link[shiny]{actionButton}.}
}
\description{
\link{messageItem} creates a message item to place in a \link{dropdownMenu}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
