% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{navbarDropdown}
\alias{navbarDropdown}
\title{Build navbar dropdown for navigation}
\usage{
navbarDropdown(text, ...)
}
\arguments{
\item{text}{Dropdown menu title.}

\item{...}{Slot for nested items such as \link{navbarTab}.}
}
\description{
This is different from \link{dropdownMenu}.
}
\keyword{internal}
