% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{notificationItem}
\alias{notificationItem}
\title{Bootstrap 4 notification item}
\usage{
notificationItem(
  text,
  icon = shiny::icon("warning"),
  status = "success",
  href = NULL,
  inputId = NULL
)
}
\arguments{
\item{text}{The notification text.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{status}{The status of the item This determines the item's background
color. Valid statuses are listed in \link{validStatuses}.}

\item{href}{An optional URL to link to.}

\item{inputId}{Whether to allow the item to act as a \link[shiny]{actionButton}.}
}
\description{
\link{messageItem} creates a message item to place in a \link{dropdownMenu}.
}
