% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{addPopover}
\alias{addPopover}
\alias{removePopover}
\title{Create a Bootstrap 4 popover from the server side}
\usage{
addPopover(
  id = NULL,
  selector = NULL,
  options,
  session = shiny::getDefaultReactiveDomain()
)

removePopover(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Popover target id.}

\item{selector}{jQuery selector. Allow more customization for the target (nested tags).}

\item{options}{List of options to pass to the popover. See \url{https://getbootstrap.com/docs/4.0/components/popovers/}.}

\item{session}{Shiny session object.}
}
\description{
\link{addPopover} adds a popover to the given target.

\link{removePopover} destroys the current targeted popover.
}
\note{
This replaces the shinyBS popover feature that is not compatible
with Bootstrap 4
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(),
    sidebar = dashboardSidebar(),
    controlbar = dashboardControlbar(),
    footer = dashboardFooter(),
    title = "Popover server",
    body = dashboardBody(
     sliderInput("obs", "Number of observations:",
       min = 0, max = 1000, value = 500
     ),
     plotOutput("distPlot")
    )
  ),
  server = function(input, output, session) {
   output$distPlot <- renderPlot({
     hist(rnorm(input$obs))
   })

   observeEvent(input$obs, {
     if (input$obs > 500) {
      addPopover(
        id = "distPlot", 
        options = list(
         content = "Vivamus sagittis lacus vel augue laoreet rutrum faucibus.",
         title = "Server popover", 
         placement = "bottom",
         trigger = "hover"
        )
      )
     } else {
       removePopover(id = "distPlot")
     }
   })
  }
 )
}
}
