% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{progressBar}
\alias{progressBar}
\title{Alias to \link{bs4ProgressBar}
See \link{bs4ProgressBar} for full details}
\usage{
progressBar(
  value,
  min = 0,
  max = 100,
  vertical = FALSE,
  striped = FALSE,
  animated = FALSE,
  status = "primary",
  size = NULL,
  label = NULL
)
}
\description{
Alias to \link{bs4ProgressBar}
See \link{bs4ProgressBar} for full details
}
