% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{blockQuote}
\alias{blockQuote}
\title{Alias to \link{bs4Quote}
See \link{bs4Quote} for full details}
\usage{
blockQuote(..., color)
}
\description{
Alias to \link{bs4Quote}
See \link{bs4Quote} for full details
}
