% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{renderInfoBox}
\alias{renderInfoBox}
\title{Alias to \link{renderbs4InfoBox}
See \link{renderbs4InfoBox} for full details}
\usage{
renderInfoBox(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link[shiny:HTML]{HTML()}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Alias to \link{renderbs4InfoBox}
See \link{renderbs4InfoBox} for full details
}
