% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skinSelector.R
\name{skinSelector}
\alias{skinSelector}
\title{AdminLTE3 skin selector}
\usage{
skinSelector()
}
\description{
This creates a skin selector element.
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     body = dashboardBody(),
     controlbar = dashboardControlbar(skinSelector(), pinned = TRUE),
     title = "Skin Selector"
   ),
   server = function(input, output) { }
 )
}
}
