% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R, R/aliases.R
\name{bs4Stars}
\alias{bs4Stars}
\alias{startBlock}
\title{AdminLTE3 stars}
\usage{
bs4Stars(maxstar = 5, grade, color = "warning")

startBlock(maxstar = 5, grade, color = "warning")
}
\arguments{
\item{maxstar}{Maximum number of stars by block.}

\item{grade}{Current score. Should be positive and lower or equal to maxstar.}

\item{color}{Star color. Valid colors are listed below:
\itemize{
 \item \code{primary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#007bff")}.
  \item \code{secondary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6c757d")}.
  \item \code{info}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#17a2b8")}.
  \item \code{success}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#28a745")}.
  \item \code{warning}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ffc107")}.
  \item \code{danger}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#dc3545")}.
}}
}
\description{
Create a block of stars (ideal for rating)
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Star example",
      starBlock(grade = 5),
      starBlock(grade = 5, color = "warning"),
      starBlock(grade = 1, color = "danger"),
      starBlock(grade = 3, color = "secondary")
     )
    ),
    title = "starBlock"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
