% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{bs4Table}
\alias{bs4Table}
\alias{bs4TableItems}
\alias{bs4TableItem}
\title{Boostrap 4 table container}
\usage{
bs4Table(
  ...,
  cardWrap = FALSE,
  headTitles,
  bordered = FALSE,
  striped = FALSE,
  width = 12
)

bs4TableItems(...)

bs4TableItem(..., dataCell = FALSE)
}
\arguments{
\item{...}{Any HTML element.}

\item{cardWrap}{Whether to wrap the table in a card. FALSE by default.}

\item{headTitles}{Table header names. Must have the same length as the number of 
\link{bs4TableItem} in \link{bs4TableItems}. Set "" to have an empty title field.}

\item{bordered}{Whether to display border between elements. FALSE by default.}

\item{striped}{Whether to displayed striped in elements. FALSE by default.}

\item{width}{Table width. 12 by default.}

\item{dataCell}{Whether the cell should be contain data or text. <td> by default.}
}
\description{
Build an argon table container

Build an bs4 table item row

Build an bs4 table item
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(), 
    sidebar = dashboardSidebar(),
    body = dashboardBody(
     bs4Table(
      cardWrap = TRUE,
      bordered = TRUE,
      striped = TRUE,
      headTitles = c(
       "PROJECT",
       "BUDGET",
       "STATUS",
       "USERS",
       "COMPLETION",
       ""
      ),
      bs4TableItems(
       bs4TableItem("bs4 Design System"),
       bs4TableItem(dataCell = TRUE, "$2,500 USD"),
       bs4TableItem(
        dataCell = TRUE, 
        dashboardBadge(
         "Pending",
         position = "right",
         color = "danger",
         rounded = TRUE
        )
       ),
       bs4TableItem(
        progressBar(value = 50)
       ),
       bs4TableItem(
        dataCell = TRUE, 
        "test"
       ),
       bs4TableItem(
        actionButton(
         "go",
         "Go"
        )
       )
      )
     )
    ), 
    footer = dashboardFooter()
  ),
  server = function(input, output) { }
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
