% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{addTooltip}
\alias{addTooltip}
\alias{removeTooltip}
\title{Create a Bootstrap 4 tooltip from the server side}
\usage{
addTooltip(
  id = NULL,
  selector = NULL,
  options,
  session = shiny::getDefaultReactiveDomain()
)

removeTooltip(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Tooltip target id.}

\item{selector}{jQuery selector. Allow more customization for the target (nested tags).}

\item{options}{List of options to pass to the tooltip. See \url{https://getbootstrap.com/docs/4.0/components/tooltips/}.}

\item{session}{Shiny session object.}
}
\description{
\link{addTooltip} adds a tooltip to the given target.

\link{removeTooltip} destroys the current targeted tooltip.
}
\note{
This replaces the shinyBS tooltip feature that is not compatible
with Bootstrap 4
}
\examples{
if (interactive()) {
 library(shiny) 
 library(bs4Dash) 
    
 shinyApp( 
   ui = dashboardPage( 
     header = dashboardHeader(), 
     sidebar = dashboardSidebar(), 
     controlbar = dashboardControlbar(), 
     footer = dashboardFooter(), 
     title = "Tooltip server", 
     body = dashboardBody( 
      sliderInput("obs", "Number of observations:", 
        min = 0, max = 1000, value = 500 
      ), 
      plotOutput("distPlot") 
     ) 
   ), 
   server = function(input, output, session) { 
     output$distPlot <- renderPlot({ 
       hist(rnorm(input$obs)) 
     }) 
         
     observeEvent(input$obs, { 
       if (input$obs > 500) { 
        addTooltip( 
          id = "distPlot",  
          options = list(
           title = "Server tooltip",  
           placement = "bottom"
          ) 
        ) 
       } else { 
         removeTooltip(id = "distPlot") 
       } 
     }) 
   } 
  )      
}
}
