% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/aliases.R
\name{updatebs4Card}
\alias{updatebs4Card}
\alias{updateCard}
\alias{updateBox}
\title{Update an AdminLTE3 card from the server side}
\usage{
updatebs4Card(
  id,
  action = c("remove", "toggle", "toggleMaximize", "restore"),
  session = shiny::getDefaultReactiveDomain()
)

updateCard(
  id,
  action = c("remove", "toggle", "toggleMaximize", "restore"),
  session = shiny::getDefaultReactiveDomain()
)

updateBox(
  id,
  action = c("remove", "toggle", "toggleMaximize", "restore"),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Card inputId.}

\item{action}{Action to trigger: \code{c("remove", "toggle", "toggleMaximize", "restore")}.}

\item{session}{Shiny session.}
}
\description{
Update an AdminLTE3 card from the server side

Alias to \link{updatebs4Card}
See \link{updatebs4Card} for full details

Alias to \link{updatebs4Card}
See \link{updatebs4Card} for full details
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     body = dashboardBody(
       actionButton(inputId = "triggerCard", label = "Trigger Card Action"),
       selectInput(
         inputId = "cardAction", 
         label = "Card action", 
         choices = c(
           "remove",
           "toggle",
           "toggleMaximize",
           "restore"
         )
       ),
       
       box(
         id = "mycard",
         title = "The plot is visible when you maximize the card", 
         closable = TRUE, 
         maximizable = TRUE,
         width = 12,
         status = "warning", 
         solidHeader = FALSE, 
         collapsible = TRUE,
         sliderInput("obs", "Number of observations:",
                     min = 0, max = 1000, value = 500
         ),
         plotOutput("distPlot")
       )
     )
   ),
   server = function(input, output, session) {
     
     output$distPlot <- renderPlot({
       if (input$mycard$maximized) {
         hist(rnorm(input$obs)) 
       }
     })
     
     observeEvent(input$triggerCard, {
       updatebs4Card(id = "mycard", session = session, action = input$cardAction)
     })
     
     observe({
       print(
         list(
           collapsed = input$mycard$collapsed,
           maximized = input$mycard$maximized,
           visible = input$mycard$visible
         )
       )
     })
   }
 )
}
}
