% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{updateCardSidebar}
\alias{updateCardSidebar}
\alias{updateBoxSidebar}
\title{Alias to \link{updatebs4CardSidebar}
See \link{updatebs4CardSidebar} for full details}
\usage{
updateCardSidebar(id, session = shiny::getDefaultReactiveDomain())

updateBoxSidebar(id, session = shiny::getDefaultReactiveDomain())
}
\description{
Alias to \link{updatebs4CardSidebar}
See \link{updatebs4CardSidebar} for full details

Alias to \link{updatebs4CardSidebar}
See \link{updatebs4CardSidebar} for full details
}
