% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{updatebs4Controlbar}
\alias{updatebs4Controlbar}
\title{Function to programmatically toggle the state of the controlbar}
\usage{
updatebs4Controlbar(inputId, session)
}
\arguments{
\item{inputId}{Controlbar id.}

\item{session}{Shiny ession object.}
}
\description{
Function to programmatically toggle the state of the controlbar
}
\examples{
if (interactive()) {

}
}
