% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{updateControlbarMenu}
\alias{updateControlbarMenu}
\title{Alias to \link{updatebs4ControlbarMenu}
See \link{updatebs4ControlbarMenu} for full details}
\usage{
updateControlbarMenu(session, inputId, selected = NULL)
}
\arguments{
\item{session}{shiny session.}

\item{inputId}{\link{bs4TabSetPanel} unique id.}

\item{selected}{the tab to be selected.}
}
\description{
Alias to \link{updatebs4ControlbarMenu}
See \link{updatebs4ControlbarMenu} for full details
}
