% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R, R/aliases.R
\name{updatebs4TabItems}
\alias{updatebs4TabItems}
\alias{updateTabItems}
\title{Change the selected sidebar tab on the client}
\usage{
updatebs4TabItems(session, inputId, selected = NULL)

updateTabItems(session, inputId, selected = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the \code{tabsetPanel}, \code{navlistPanel},
or \code{navbarPage} object.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}
}
\description{
This function controls the active tab of \code{\link{tabItems}} from the
server. It behaves just like \code{\link[shiny]{updateTabsetPanel}}.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(skin = "dark"),
     body = dashboardBody(
       tabItems(
         tabItem(
           tabName = "tab1",
           sliderInput("obs", "Number of observations:",
                       min = 0, max = 1000, value = 500
           ),
           plotOutput("distPlot")
         ),
         tabItem(
           tabName = "tab2",
           checkboxGroupInput("variable", "Variables to show:",
                              c("Cylinders" = "cyl",
                                "Transmission" = "am",
                                "Gears" = "gear")),
           tableOutput("data")
         ),
         tabItem(
           tabName = "tab3",
           checkboxInput("val", "Some value", FALSE),
           textOutput("value")
         ),
         tabItem(
           tabName = "tab4",
           "Nothing special here!"
         ),
         tabItem(
           tabName = "tab5",
           "Tab 5"
         ),
         tabItem(
           tabName = "tab6",
           "Tab 6"
         ),
         tabItem(
           tabName = "tab7",
           "Tab 7"
         )
       )
     ),
     sidebar = dashboardSidebar(
       skin = "light",
       inputId = "sidebarState",
       sidebarMenu(
         id = "sidebar",
         menuItem(
           text = "Tab 1",
           tabName = "tab1",
           icon = icon("shuttle-van")
         ),
         menuItem(
           text = "Tab 2",
           tabName = "tab2",
           icon = icon("space-shuttle"),
           selected = TRUE
         ),
         menuItem(
           text = "Item List 1",
           icon = icon("bars"),
           startExpanded = TRUE,
           menuSubItem(
             text = "Item 3",
             tabName = "tab3",
             icon = icon("circle-thin")
           ),
           menuSubItem(
             text = "Item 4",
             tabName = "tab4",
             icon = icon("circle-thin")
           )
         ),
         menuItem(
           text = "Item List 2",
           icon = icon("bars"),
           startExpanded = FALSE,
           menuSubItem(
             text = "Item 5",
             tabName = "tab5",
             icon = icon("circle-thin")
           ),
           menuSubItem(
             text = "Item 6",
             tabName = "tab6",
             icon = icon("circle-thin")
           )
         ),
         menuItem(
           text = "Tab 7",
           tabName = "tab7",
           icon = icon("home")
         )
       )
     ),
     controlbar = dashboardControlbar(
       skin = "light",
       sliderInput(
         inputId = "controller",
         label = "Update the first tabset",
         min = 1,
         max = 6,
         value = 2
       )
     ),
     footer = bs4DashFooter()
   ),
   server = function(input, output, session) {
     observe(print(input$sidebarItemExpanded))
     observe(print(input$sidebar))
     
     # update tabset1
     observeEvent(input$controller, {
       updateTabItems(
         session, 
         inputId = "sidebar", 
         selected = paste0("tab", input$controller)
       )
     }, ignoreInit = TRUE)
     
     output$distPlot <- renderPlot({
       hist(rnorm(input$obs))
     })
     
     output$data <- renderTable({
       mtcars[, c("mpg", input$variable), drop = FALSE]
     }, rownames = TRUE)
     
     output$value <- renderText({ input$val })
     
   }
 )
}
}
