% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{updateControlbar}
\alias{updateControlbar}
\title{Function to programmatically toggle the state of the controlbar}
\usage{
updateControlbar(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Controlbar id.}

\item{session}{Shiny session object.}
}
\description{
Function to programmatically toggle the state of the controlbar
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     body = dashboardBody(
       actionButton(inputId = "controlbarToggle", label = "Toggle Controlbar")
     ),
     controlbar = dashboardControlbar(
       id = "controlbar",
       collapsed = FALSE,
       overlay = TRUE
     )
   ),
   server = function(input, output, session) {
     
     observeEvent(input$controlbar, {
       if (input$controlbar) {
         showModal(modalDialog(
           title = "Alert",
           "The controlbar is opened.",
           easyClose = TRUE,
           footer = NULL
         ))
       }
     })
     
     observeEvent(input$controlbarToggle, {
       updateControlbar(id = "controlbar", session = session)
     })
     
     observe({
       print(input$controlbar)
     })
   }
 )
}
}
