% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userMessage}
\alias{userMessage}
\title{AdminLTE3 user message}
\usage{
userMessage(..., author = NULL, date = NULL, src = NULL,
  side = c(NULL, "right"))
}
\arguments{
\item{...}{Message text.}

\item{author}{Message author.}

\item{date}{Message date.}

\item{src}{Message author image path or url.}

\item{side}{Side where author is displayed: NULL (left) or "right".}
}
\description{
Create a user message
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
