% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userPostTagItem}
\alias{userPostTagItem}
\title{AdminLTE3 user post tool item}
\usage{
userPostTagItem(...)
}
\arguments{
\item{...}{Tool content such as label, button, ...}
}
\description{
Create a user post tool item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
