% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userPostTagItems}
\alias{userPostTagItems}
\title{AdminLTE3 user post tool item container}
\usage{
userPostTagItems(...)
}
\arguments{
\item{...}{Slot for \link{userPostTagItem}.}
}
\description{
Create a user post tool item container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
