% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateColors}
\alias{validateColors}
\title{Validate all colors & Bootstrap statuses}
\usage{
validateColors(color)
}
\arguments{
\item{color}{\code{chr} string with status name or color name}
}
\value{
\code{lgl}
}
\description{
Validate all colors & Bootstrap statuses
}
\seealso{
grDevices::color
}
