% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{validColors}
\alias{validColors}
\title{Valid colors}
\description{
These are valid colors for various dashboard components. Valid colors are
listed below:
}
\details{
\itemize{
  \item \code{indigo} Indigo
  \item \code{purple} Purple
  \item \code{pink} Pink
  \item \code{orange} Orange
  \item \code{teal} Blue/Green
}
}
\keyword{internal}
