% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{validNuances}
\alias{validNuances}
\title{Valid nuances}
\description{
These nuances strings correspond to colors as defined in AdminLTE's CSS.
Although the colors can vary depending on the particular CSS selector, they
generally appear as follows:
}
\details{
\itemize{
\item \code{gray-dark} Gray dark
\item \code{gray} Gray
\item \code{white} White
}
}
\keyword{internal}
