% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{boxLayout}
\alias{boxLayout}
\title{Alias to \link{bs4CardLayout}
See \link{bs4CardLayout} for full details}
\usage{
boxLayout(..., type = c("group", "deck", "columns"))
}
\description{
Alias to \link{bs4CardLayout}
See \link{bs4CardLayout} for full details
}
