% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/aliases.R
\name{bs4CardLabel}
\alias{bs4CardLabel}
\alias{cardLabel}
\alias{boxLabel}
\title{Create a label for Boostrap 4 card}
\usage{
bs4CardLabel(text, status, tooltip = NULL)

cardLabel(text, status, tooltip = NULL)

boxLabel(text, status, tooltip = NULL)
}
\arguments{
\item{text}{Label text. In practice only few letters or a number.}

\item{status}{label color status. See \link{getAdminLTEColors}.}

\item{tooltip}{Label tooltip text on hover.}
}
\description{
Create a label for Boostrap 4 card

Alias to \link{bs4CardLabel}
See \link{bs4CardLabel} for full details

Alias to \link{bs4CardLabel}
See \link{bs4CardLabel} for full details
}
\seealso{
Other boxWidgets: 
\code{\link{attachmentBlock}()},
\code{\link{bs4CardSidebar}()},
\code{\link{bs4Carousel}()},
\code{\link{bs4SocialCard}()},
\code{\link{bs4Timeline}()},
\code{\link{cardDropdown}()},
\code{\link{cardProfile}()},
\code{\link{descriptionBlock}()},
\code{\link{userPost}()}
}
\concept{boxWidgets}
