% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{cardSidebar}
\alias{cardSidebar}
\alias{boxSidebar}
\title{Alias to \link{bs4CardSidebar}
See \link{bs4CardSidebar} for full details}
\usage{
cardSidebar(
  ...,
  id = NULL,
  width = "25\%",
  background = "#333a40",
  startOpen = FALSE,
  icon = shiny::icon("cogs")
)

boxSidebar(
  ...,
  id = NULL,
  width = "25\%",
  background = "#333a40",
  startOpen = FALSE,
  icon = shiny::icon("cogs")
)
}
\description{
Alias to \link{bs4CardSidebar}
See \link{bs4CardSidebar} for full details

Alias to \link{bs4CardSidebar}
See \link{bs4CardSidebar} for full details
}
