% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/aliases.R
\name{bs4CardSidebar}
\alias{bs4CardSidebar}
\alias{cardSidebar}
\title{Create a sidebar for Boostrap 4 card}
\usage{
bs4CardSidebar(
  ...,
  inputId = NULL,
  width = "25\%",
  background = "#333a40",
  startOpen = FALSE,
  icon = "cogs"
)

cardSidebar(
  ...,
  inputId = NULL,
  width = "25\%",
  background = "#333a40",
  startOpen = FALSE,
  icon = "cogs"
)
}
\arguments{
\item{...}{Sidebar content.}

\item{inputId}{Unique sidebar id. Useful if you want to use \link{updatebs4CardSidebar}.}

\item{width}{Sidebar width in percentage. 25\% by default. A character value of any width CSS understands (e.g. "100px").}

\item{background}{Sidebar background color. Dark by default.}

\item{startOpen}{Whether the sidebar is open at start. FALSE by default.}

\item{icon}{Sidebar icon.}
}
\description{
To insert in the sidebar slot of \link{bs4Card}.
}
