% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R, R/aliases.R
\name{bs4DashBody}
\alias{bs4DashBody}
\alias{dashboardBody}
\title{Create a Boostrap 4 dashboard body}
\usage{
bs4DashBody(...)

dashboardBody(...)
}
\arguments{
\item{...}{Body content, slot for \link{bs4TabItems}.}
}
\description{
Build an adminLTE3 dashboard body
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
