% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R, R/aliases.R
\name{bs4DashBrand}
\alias{bs4DashBrand}
\alias{dashboardBrand}
\title{Alternative to simple text title}
\usage{
bs4DashBrand(title, color = NULL, src = NULL, image = NULL, opacity = 0.8)

dashboardBrand(title, color = NULL, src = NULL, image = NULL, opacity = 0.8)
}
\arguments{
\item{title}{Brand title.}

\item{color}{Brand color.}

\item{src}{External link to point to.}

\item{image}{External image location.}

\item{opacity}{Brand opacity: value between 0 and 1.}
}
\value{
A title tag to be inserted in the title slot of \link{bs4DashNavbar}.
}
\description{
Alternative to simple text title

Alias to \link{bs4DashBrand}
See \link{bs4DashBrand} for full details
}
