% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{bs4DashControlbar}
\alias{bs4DashControlbar}
\title{Create a Boostrap 4 dashboard right sidebar}
\usage{
bs4DashControlbar(..., disable = FALSE, skin = "dark", title = NULL,
  width = 250)
}
\arguments{
\item{...}{Any UI element.}

\item{disable}{If \code{TRUE}, the sidebar will be disabled.}

\item{skin}{Controlbar skin. "dark" or "light".}

\item{title}{Controlbar title.}

\item{width}{Controlbar width. 250 px by default.}
}
\description{
Build an adminLTE3 dashboard right sidebar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
