% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{controlbarItem}
\alias{controlbarItem}
\title{Alias to \link{bs4DashControlbarItem}
See \link{bs4DashControlbarItem} for full details}
\usage{
controlbarItem(tabName, ..., active = FALSE)
}
\arguments{
\item{tabName}{Tab name: it will be also passed as the id argument. Must be unique.}

\item{...}{Tab content.}

\item{active}{Whether the tab is active or not. FALSE bu default.}
}
\description{
Alias to \link{bs4DashControlbarItem}
See \link{bs4DashControlbarItem} for full details
}
