% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{controlbarMenu}
\alias{controlbarMenu}
\title{Alias to \link{bs4DashControlbarMenu}
See \link{bs4DashControlbarMenu} for full details}
\usage{
controlbarMenu(
  ...,
  id = NULL,
  side = "left",
  status = NULL,
  tabStatus = NULL,
  .list = NULL,
  vertical = FALSE,
  type = NULL
)
}
\arguments{
\item{...}{Slot for \link{bs4TabPanel}.}

\item{id}{Unique \link{bs4TabSetPanel} id. NULL by default. Set a value
to get the currently selected tab.}

\item{side}{Side of the box the tabs should be on (\code{"left"} or
\code{"right"}). Default to "left".}

\item{status}{The status of the item This determines the item's background
  color.  Valid statuses are defined as follows:
\itemize{

  \item \code{primary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#007bff")}.
  \item \code{secondary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6c757d")}.
  \item \code{info}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#17a2b8")}.
  \item \code{success}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#28a745")}.
  \item \code{warning}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ffc107")}.
  \item \code{danger}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#dc3545")}.
  \item \code{black}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#000000")}.
  \item \code{gray-dark}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#343a40")}.
  \item \code{gray}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#adb5bd")}.
  \item \code{light}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#1f2d3d")}.
  \item \code{indigo}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6610f2")}.
  \item \code{lightblue}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3c8dbc")}.
  \item \code{navy}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#001f3f")}.
  \item \code{purple}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#605ca8")}.
  \item \code{fuchsia}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#f012be")}.
  \item \code{pink}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#e83e8c")}.
  \item \code{maroon}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#d81b60")}.
  \item \code{orange}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ff851b")}.
  \item \code{lime}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#01ff70")}.
  \item \code{teal}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#39cccc")}.
  \item \code{olive}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3d9970")}.
}
Only primary, success, info, warning and danger are compatible with solidHeader!}

\item{tabStatus}{The status of the tabs buttons over header. "primary", "secondary", "success", "warning", "danger", "white", "light", "dark", "transparent".
NULL by default, "light" if status is set.   
A vector is possible with a colour for each tab button}

\item{.list}{When elements are programmatically added, pass them here instead of in ...}

\item{vertical}{Whether to display tabs in a vertical mode. FALSE by default.}

\item{type}{TabPanel type: "tabs" or "pills". "pills" is the default if type is NULL.}
}
\description{
Alias to \link{bs4DashControlbarMenu}
See \link{bs4DashControlbarMenu} for full details
}
