% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{controlbarMenu}
\alias{controlbarMenu}
\title{Alias to \link{bs4DashControlbarMenu}
See \link{bs4DashControlbarMenu} for full details}
\usage{
controlbarMenu(
  ...,
  id = NULL,
  side = "left",
  status = NULL,
  tabStatus = NULL,
  .list = NULL,
  vertical = FALSE,
  type = NULL
)
}
\arguments{
\item{...}{Slot for \link{bs4TabPanel}.}

\item{id}{Unique \link{bs4TabSetPanel} id. NULL by default. Set a value
to get the currently selected tab.}

\item{side}{Side of the box the tabs should be on (\code{"left"} or
\code{"right"}). Default to "left".}

\item{status}{The status of the card header. "primary", "secondary", "success", "warning", "danger", "white", "light", "dark", "transparent". NULL by default.}

\item{tabStatus}{The status of the tabs buttons over header. "primary", "secondary", "success", "warning", "danger", "white", "light", "dark", "transparent".
NULL by default, "light" if status is set.   
A vector is possible with a colour for each tab button}

\item{.list}{When elements are programmatically added, pass them here instead of in ...}

\item{vertical}{Whether to display tabs in a vertical mode. FALSE by default.}

\item{type}{TabPanel type: "tabs" or "pills". "pills" is the default if type is NULL.}
}
\description{
Alias to \link{bs4DashControlbarMenu}
See \link{bs4DashControlbarMenu} for full details
}
