% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R, R/aliases.R
\name{bs4DashFooter}
\alias{bs4DashFooter}
\alias{dashboardFooter}
\title{Create a Boostrap 4 dashboard footer}
\usage{
bs4DashFooter(..., copyrights = NULL, right_text = NULL, fixed = FALSE)

dashboardFooter(..., copyrights = NULL, right_text = NULL, fixed = FALSE)
}
\arguments{
\item{...}{Left text.}

\item{copyrights}{Copyrights, if any.}

\item{right_text}{Right text.}

\item{fixed}{Whether to fix the footer. Default to FALSE.}
}
\description{
Build an adminLTE3 dashboard footer
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
