% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{bs4DashNavbar}
\alias{bs4DashNavbar}
\title{Create a Boostrap 4 dashboard navbar}
\usage{
bs4DashNavbar(..., skin = "light", status = "white", border = TRUE,
  sidebarIcon = "bars", controlbarIcon = "th-large", leftUi = NULL,
  rightUi = NULL)
}
\arguments{
\item{...}{Any UI element between left and right Ui.}

\item{skin}{Navbar skin. "dark" or "light".}

\item{status}{Navbar status. "white" by default: "primary", "danger", "warning",
"success", "info", "white" or "gray-light".}

\item{border}{Whether to separate the navbar and body by a border. TRUE by default.}

\item{sidebarIcon}{Icon of the main sidebar toggle.}

\item{controlbarIcon}{Icon to toggle the controlbar (left).}

\item{leftUi}{Custom left Ui content. Any Ui element.}

\item{rightUi}{Custom right Ui content. Any Ui element.}
}
\description{
Build an adminLTE3 dashboard navbar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
