% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardPage.R
\name{bs4DashPage}
\alias{bs4DashPage}
\title{Create a Boostrap 4 dashboard page}
\usage{
bs4DashPage(navbar = NULL, sidebar = NULL, body = NULL,
  controlbar = NULL, footer = NULL, title = NULL, old_school = FALSE)
}
\arguments{
\item{navbar}{Bootstrap 4 dashboard navbar.}

\item{sidebar}{Bootstrap 4 dashboard main sidebar.}

\item{body}{Bootstrap 4 dashboard body wrapper.}

\item{controlbar}{Bootstrap 4 dashboard control sidebar (right side).}

\item{footer}{Bootstrap 4 dashboard footer.}

\item{title}{App title.}

\item{old_school}{Whether to use the wonderful sketchy design for BS4. FALSE
by default.}
}
\description{
Build an adminLTE3 dashboard page
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)

 shiny::shinyApp(
   ui = bs4DashPage(
    navbar = bs4DashNavbar(),
    sidebar = bs4DashSidebar(),
    controlbar = bs4DashControlbar(),
    footer = bs4DashFooter(),
    title = "test",
    body = bs4DashBody()
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
