% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardPage.R
\name{bs4DashPage}
\alias{bs4DashPage}
\title{Create a Boostrap 4 dashboard page}
\usage{
bs4DashPage(navbar = NULL, sidebar = NULL, body = NULL,
  controlbar = NULL, footer = NULL, title = NULL,
  old_school = FALSE, sidebar_collapsed = FALSE,
  controlbar_collapsed = TRUE)
}
\arguments{
\item{navbar}{Slot for \link{bs4DashNavbar}.}

\item{sidebar}{Slot for \link{bs4DashSidebar}.}

\item{body}{Slot for \link{bs4DashBody}.}

\item{controlbar}{Slot for \link{bs4DashControlbar} (right side).}

\item{footer}{Slot for \link{bs4DashFooter}.}

\item{title}{App title.}

\item{old_school}{Whether to use the wonderful sketchy design for Bootstrap 4. FALSE
by default.}

\item{sidebar_collapsed}{Whether the sidebar is collapsed of not at start. FALSE by default.}

\item{controlbar_collapsed}{Whether the sidebar is collapsed of not at start. TRUE by default.}
}
\description{
Build an adminLTE3 dashboard page
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)

 shiny::shinyApp(
   ui = bs4DashPage(
    navbar = bs4DashNavbar(),
    sidebar = bs4DashSidebar(),
    controlbar = bs4DashControlbar(),
    footer = bs4DashFooter(),
    title = "test",
    body = bs4DashBody()
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
