% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardPage.R, R/aliases.R
\name{bs4DashPage}
\alias{bs4DashPage}
\alias{dashboardPage}
\title{Create a Boostrap 4 dashboard page}
\usage{
bs4DashPage(
  header,
  sidebar,
  body,
  controlbar = NULL,
  footer = NULL,
  title = NULL,
  freshTheme = NULL,
  preloader = NULL,
  options = NULL
)

dashboardPage(
  header,
  sidebar,
  body,
  controlbar = NULL,
  footer = NULL,
  title = NULL,
  freshTheme = NULL,
  preloader = NULL,
  options = NULL
)
}
\arguments{
\item{header}{Slot for \link{bs4DashNavbar}.}

\item{sidebar}{Slot for \link{bs4DashSidebar}.}

\item{body}{Slot for \link{bs4DashBody}.}

\item{controlbar}{Slot for \link{bs4DashControlbar} (right side).}

\item{footer}{Slot for \link{bs4DashFooter}.}

\item{title}{App title.}

\item{freshTheme}{A skin powered by the fresh package. Not compatible with skin.
See \url{https://dreamrs.github.io/fresh/articles/vars-shinydashboard.html}.}

\item{preloader}{bs4Dash uses waiter (see \url{https://waiter.john-coene.com/#/}).
Pass a nested list with 2 elements like \code{list(waiter = list(html = spin_1(), color = "#333e48"), duration = 5)}.
\code{waiter} expects to provide a sub-list to configure \link[waiter]{waiter_show_on_load} (refer to
the package help for all styles). \code{duration} defines the loader timeout.}

\item{options}{Extra option to overwrite the vanilla AdminLTE configuration. See 
\url{https://adminlte.io/themes/AdminLTE/documentation/index.html#adminlte-options}.
Expect a list.
\url{https://www.w3schools.com/cssref/css_colors.asp}.}
}
\description{
Build an adminLTE3 dashboard page
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 library(fresh)

 shinyApp(
   ui = dashboardPage(
     freshTheme = create_theme(
      bs4dash_vars(
        navbar_light_color = "#bec5cb",
        navbar_light_active_color = "#FFF",
        navbar_light_hover_color = "#FFF"
      ),
      bs4dash_yiq(
        contrasted_threshold = 10,
        text_dark = "#FFF", 
        text_light = "#272c30"
      ),
      bs4dash_layout(
        main_bg = "#353c42"
      ),
      bs4dash_sidebar_light(
        bg = "#272c30", 
        color = "#bec5cb",
        hover_color = "#FFF",
        submenu_bg = "#272c30", 
        submenu_color = "#FFF", 
        submenu_hover_color = "#FFF"
      ),
      bs4dash_status(
        primary = "#5E81AC", danger = "#BF616A", light = "#272c30"
      ),
      bs4dash_color(
        gray_900 = "#FFF", white = "#272c30"
      )
     ),
     options = NULL,
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     body = dashboardBody(
      box(status = "danger"),
      box(status = "primary"),
      box(status = "orange")
     ),
     controlbar = dashboardControlbar(),
     title = "DashboardPage"
   ),
   server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
