% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R, R/aliases.R
\name{bs4DashSidebar}
\alias{bs4DashSidebar}
\alias{dashboardSidebar}
\title{Create a Boostrap 4 dashboard main sidebar}
\usage{
bs4DashSidebar(
  ...,
  inputId = NULL,
  disable = FALSE,
  title = NULL,
  skin = "dark",
  status = "primary",
  brandColor = NULL,
  url = NULL,
  src = NULL,
  elevation = 4,
  opacity = 0.8,
  expand_on_hover = TRUE,
  fixed = TRUE
)

dashboardSidebar(
  ...,
  inputId = NULL,
  disable = FALSE,
  title = NULL,
  skin = "dark",
  status = "primary",
  brandColor = NULL,
  url = NULL,
  src = NULL,
  elevation = 4,
  opacity = 0.8,
  expand_on_hover = TRUE,
  fixed = TRUE
)
}
\arguments{
\item{...}{Slot for \link{bs4SidebarMenu}.}

\item{inputId}{Recover the state of the sidebar.}

\item{disable}{If \code{TRUE}, the sidebar will be disabled.}

\item{title}{Sidebar title.}

\item{skin}{Sidebar skin. "dark" or "light"}

\item{status}{Sidebar status. "primary", "danger", "warning", "success", "info".}

\item{brandColor}{Brand color. NULL by default: "primary", "danger", "warning",
"success", "info", "white" or "gray-light".}

\item{url}{Sidebar brand link.}

\item{src}{Sidebar brand image.}

\item{elevation}{Sidebar elevation. 4 by default (until 5).}

\item{opacity}{Sidebar brand opacity. From 0 to 1. 0.8 by default.}

\item{expand_on_hover}{Whether to expand the sidebar om hover. TRUE by default.}

\item{fixed}{Whether to fix the sidebar. Default to TRUE.}
}
\description{
Build an adminLTE3 dashboard main sidebar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
