% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R, R/aliases.R
\name{bs4DropdownMenu}
\alias{bs4DropdownMenu}
\alias{dropdownMenu}
\title{Create a Boostrap 4 dashboard dropdown menu}
\usage{
bs4DropdownMenu(
  ...,
  show = FALSE,
  labelText = NULL,
  src = NULL,
  status = c("primary", "warning", "danger", "info", "success"),
  menuIcon = "bell",
  align = "right"
)

dropdownMenu(
  ...,
  show = FALSE,
  labelText = NULL,
  src = NULL,
  status = c("primary", "warning", "danger", "info", "success"),
  menuIcon = "bell",
  align = "right"
)
}
\arguments{
\item{...}{Slot for \link{bs4DropdownMenuItem}.}

\item{show}{Whether to start with the dropdown open. FALSE by default.}

\item{labelText}{Dropdown label text.}

\item{src}{Dropdown link to an external ressource.}

\item{status}{Dropdown menu status. "primary", "success", "warning", "danger" or "info".}

\item{menuIcon}{Fontawesome icon (default = "bell")}

\item{align}{Menu alignment (default = "right")}
}
\description{
Build an adminLTE3 dashboard dropdown menu
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shiny::shinyApp(
   ui = bs4DashPage(
     navbar = bs4DashNavbar(
      rightUi = bs4DropdownMenu(
       show = FALSE,
       status = "danger",
       src = "https://www.google.fr",
       bs4DropdownMenuItem(
         inputId = "triggerAction",
         message = "message 1",
         from = "Divad Nojnarg",
         src = "https://adminlte.io/themes/v3/dist/img/user3-128x128.jpg",
         time = "today",
         status = "danger",
         type = "message"
       ),
       bs4DropdownMenuItem(
         message = "message 2",
         from = "Nono Gueye",
         src = "https://adminlte.io/themes/v3/dist/img/user3-128x128.jpg",
         time = "yesterday",
         status = "success",
         type = "message"
       )
      )
     ),
     sidebar = bs4DashSidebar(),
     controlbar = bs4DashControlbar(),
     footer = bs4DashFooter(),
     title = "test",
     body = bs4DashBody()
   ),
   server = function(input, output) {
    observeEvent(input$triggerAction, {
     showModal(modalDialog(
      title = "Important message",
      "This is an important message!"
     ))
    })
   }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
