% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{bs4DropdownMenuItem}
\alias{bs4DropdownMenuItem}
\title{Create a Boostrap 4 dashboard dropdown menu item}
\usage{
bs4DropdownMenuItem(text, date = NULL, icon = "envelope")
}
\arguments{
\item{text}{Item content.}

\item{date}{Item date.}

\item{icon}{Item icon.}
}
\description{
Build an adminLTE3 dashboard dropdown menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
