% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R, R/aliases.R
\name{bs4DropdownMenuItem}
\alias{bs4DropdownMenuItem}
\alias{dropdownMenuItem}
\title{Create a Boostrap 4 dashboard dropdown menu item}
\usage{
bs4DropdownMenuItem(
  inputId = NULL,
  message,
  from = NULL,
  time = NULL,
  icon = "info-circle",
  src = NULL,
  status = NULL,
  type = c("notification", "message")
)

dropdownMenuItem(
  inputId = NULL,
  message,
  from = NULL,
  time = NULL,
  icon = "info-circle",
  src = NULL,
  status = NULL,
  type = c("notification", "message")
)
}
\arguments{
\item{inputId}{Whether to allow the item to act as a \link[shiny]{actionButton}.}

\item{message}{Item content.}

\item{from}{Item sender. Only if type is message.}

\item{time}{Item date.}

\item{icon}{Item icon.}

\item{src}{Item image. Only if type is message.}

\item{status}{Item status. Leave NULL if type is notification. See
\link{getAdminLTEColors} for valid statuses.
If not NULL, Indicate the message priority.}

\item{type}{Item type: notification or message.}
}
\description{
Build an adminLTE3 dashboard dropdown menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
