% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{bs4PopoverServer}
\alias{bs4PopoverServer}
\title{Create a Bootstrap 4 Popover from the server side}
\usage{
bs4PopoverServer(
  target,
  content,
  title,
  placement = c("top", "bottom", "left", "right"),
  session
)
}
\arguments{
\item{target}{Popover target.}

\item{content}{Popover content.}

\item{title}{Popover title.}

\item{placement}{Popover placement: "top", "bottom", "left" or "right".}

\item{session}{Shiny session object.}
}
\description{
This replaces the shinyBS popover feature that is not compatible
with Bootstrap 4
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shiny::shinyApp(
  ui = bs4DashPage(
    enable_preloader = TRUE,
    navbar = bs4DashNavbar(),
    sidebar = bs4DashSidebar(),
    controlbar = bs4DashControlbar(),
    footer = bs4DashFooter(),
    title = "test",
    body = bs4DashBody(
     actionButton("goButton", "Show popover!"),
     actionButton("goButton2", "You can't see me first!")
    )
  ),
  server = function(input, output, session) {
   observeEvent(input$goButton, {
      bs4PopoverServer(
        session = session, 
        target = "goButton2", 
        content = "Vivamus sagittis lacus vel augue laoreet rutrum faucibus.",
        title = "Server popover", 
        placement = "bottom"
      )
    })
  }
 )
}
}
