% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{bs4SidebarMenu}
\alias{bs4SidebarMenu}
\title{Create a Boostrap 4 dashboard main sidebar menu}
\usage{
bs4SidebarMenu(...)
}
\arguments{
\item{...}{Slot for \link{bs4SidebarMenuItem} or \link{bs4SidebarHeader}.}
}
\description{
Build an adminLTE3 dashboard main sidebar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
