% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{bs4SidebarMenuItem}
\alias{bs4SidebarMenuItem}
\title{Create a Boostrap 4 dashboard main sidebar menu item}
\usage{
bs4SidebarMenuItem(text, ..., tabName = NULL, icon = NULL,
  startExpanded = FALSE)
}
\arguments{
\item{text}{Item name.}

\item{...}{\link{bs4SidebarMenuSubItem}.}

\item{tabName}{Should correspond exactly to the tabName given in \code{\link{bs4TabItem}}.}

\item{icon}{Item icon.}

\item{startExpanded}{Whether to expand the \link{bs4SidebarMenuItem} at start.}
}
\description{
Build an adminLTE3 dashboard main sidebar menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
