% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R, R/aliases.R
\name{bs4SidebarUserPanel}
\alias{bs4SidebarUserPanel}
\alias{sidebarUserPanel}
\title{Create a Boostrap 4 dashboard main sidebar user panel}
\usage{
bs4SidebarUserPanel(img = NULL, text = NULL)

sidebarUserPanel(img = NULL, text = NULL)
}
\arguments{
\item{img}{User panel image path or url.}

\item{text}{User panel text.}
}
\description{
Build an adminLTE3 dashboard main sidebar user panel
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
