% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R, R/aliases.R
\name{bs4TabItem}
\alias{bs4TabItem}
\alias{tabItem}
\title{One tab to put inside a tab items container}
\usage{
bs4TabItem(tabName = NULL, ...)

tabItem(tabName = NULL, ...)
}
\arguments{
\item{tabName}{The name of a tab. This must correspond to the \code{tabName}
of a \code{\link{bs4SidebarMenuItem}}.}

\item{...}{Contents of the tab.}
}
\description{
One tab to put inside a tab items container

Alias to \link{bs4TabItem}
See \link{bs4TabItem} for full details
}
