% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R, R/aliases.R
\name{bs4TabItems}
\alias{bs4TabItems}
\alias{tabItems}
\title{A container for tab items}
\usage{
bs4TabItems(...)

tabItems(...)
}
\arguments{
\item{...}{Items to put in the container. Each item should be a
\code{\link{bs4TabItem}}.}
}
\description{
A container for tab items

Alias to \link{bs4TabItems}
See \link{bs4TabItems} for full details
}
