% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/dashboardControlbar.R, R/aliases.R
\name{bs4TabPanel}
\alias{bs4TabPanel}
\alias{bs4DashControlbarItem}
\alias{tabPanel}
\title{Create a tabPanel}
\usage{
bs4TabPanel(..., tabName, active = FALSE)

bs4DashControlbarItem(..., tabName, active = FALSE)

tabPanel(..., tabName, active = FALSE)
}
\arguments{
\item{...}{Tab content}

\item{tabName}{Tab name: it will be also passed as the id argument. Should be unique.}

\item{active}{Whether the tab is active or not. FALSE bu default.}
}
\description{
To be included in a bs4TabCard
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
