% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{bs4TabSetPanel}
\alias{bs4TabSetPanel}
\title{Create a tabSetPanel}
\usage{
bs4TabSetPanel(..., side)
}
\arguments{
\item{...}{Slot for \link{bs4TabPanel}.}

\item{side}{Side of the box the tabs should be on (\code{"left"} or
\code{"right"}).}
}
\description{
Imported by bs4TabCard. Do not use outside!
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
