% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/dashboardControlbar.R, R/aliases.R
\name{bs4TabSetPanel}
\alias{bs4TabSetPanel}
\alias{bs4DashControlbarMenu}
\alias{tabsetPanel}
\title{Create a tabSetPanel}
\usage{
bs4TabSetPanel(
  ...,
  id = NULL,
  side = "left",
  status = NULL,
  tabStatus = NULL,
  .list = NULL,
  vertical = FALSE,
  type = NULL
)

bs4DashControlbarMenu(
  ...,
  id = NULL,
  side = "left",
  status = NULL,
  tabStatus = NULL,
  .list = NULL,
  vertical = FALSE,
  type = NULL
)

tabsetPanel(
  ...,
  id = NULL,
  side = "left",
  status = NULL,
  tabStatus = NULL,
  .list = NULL,
  vertical = FALSE,
  type = NULL
)
}
\arguments{
\item{...}{Slot for \link{bs4TabPanel}.}

\item{id}{Unique \link{bs4TabSetPanel} id. NULL by default. Set a value
to get the currently selected tab.}

\item{side}{Side of the box the tabs should be on (\code{"left"} or
\code{"right"}). Default to "left".}

\item{status}{The status of the card header. "primary", "secondary", "success", "warning", "danger", "white", "light", "dark", "transparent". NULL by default.}

\item{tabStatus}{The status of the tabs buttons over header. "primary", "secondary", "success", "warning", "danger", "white", "light", "dark", "transparent".
NULL by default, "light" if status is set.   
A vector is possible with a colour for each tab button}

\item{.list}{When elements are programmatically added, pass them here instead of in ...}

\item{vertical}{Whether to display tabs in a vertical mode. FALSE by default.}

\item{type}{TabPanel type: "tabs" or "pills". "pills" is the default if type is NULL.}
}
\description{
Imported by \link{bs4TabCard} but can be used alone.
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)

 shiny::shinyApp(
   ui = bs4DashPage(
    navbar = bs4DashNavbar(),
    sidebar = bs4DashSidebar(),
    controlbar = bs4DashControlbar(),
    footer = bs4DashFooter(),
    title = "test",
    body = bs4DashBody(
     
     # manually inserted panels
     bs4TabSetPanel(
      id = "tabcard",
      side = "left",
      bs4TabPanel(
       tabName = "Tab 1", 
       active = FALSE,
       "Content 1"
      ),
      bs4TabPanel(
       tabName = "Tab 2", 
       active = TRUE,
       "Content 2"
      ),
      bs4TabPanel(
       tabName = "Tab 3", 
       active = FALSE,
       "Content 3"
      )
     ),
     
     br(), br(),
     # programmatically inserted panels
     bs4TabSetPanel(
       id = "tabset",
       side = "left",
       .list = lapply(1:3, function(i) {
         bs4TabPanel(
           tabName = paste0("Tab", i), 
           active = FALSE,
           paste("Content", i)
         )
       })
      ),
      
      br(), br(),
      # vertical tabset
      bs4TabSetPanel(
       id = "verttabset",
       side = "left",
       vertical = TRUE,
       .list = lapply(1:3, function(i) {
         bs4TabPanel(
           tabName = paste0("Tab", i), 
           active = FALSE,
           paste("Content", i)
         )
       })
      )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
