% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbacks.R
\name{bs4TooltipUI}
\alias{bs4TooltipUI}
\title{Create a Bootstrap 4 Tooltip from the UI side}
\usage{
bs4TooltipUI(tag, title, placement = c("top", "bottom", "left", "right"))
}
\arguments{
\item{tag}{Tooltip target.}

\item{title}{Tooltip title.}

\item{placement}{Tooltipe placement: "top", "bottom", "left" or "right".}
}
\description{
This replaces the shinyBS tooltip feature that is not compatible
with Bootstrap 4
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shiny::shinyApp(
  ui = bs4DashPage(
    enable_preloader = TRUE,
    navbar = bs4DashNavbar(),
    sidebar = bs4DashSidebar(),
    controlbar = bs4DashControlbar(),
    footer = bs4DashFooter(),
    title = "test",
    body = bs4DashBody(
     bs4TooltipUI(
      actionButton("goButton", "Hover to see the tooltip"),
      title = "My tooltip",
      placement = "top"
     )
    )
  ),
  server = function(input, output) {}
 )
}
}
