% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R, R/aliases.R
\name{cardDropdownItem}
\alias{cardDropdownItem}
\alias{boxDropdownItem}
\title{Create a box dropdown item}
\usage{
cardDropdownItem(..., id = NULL, href = NULL, icon = NULL)

boxDropdownItem(..., id = NULL, href = NULL, icon = NULL)
}
\arguments{
\item{...}{Item content.}

\item{id}{If passed, the item will behave like an action button.}

\item{href}{Target url or page.}

\item{icon}{Optional icon. Expect \link[shiny]{icon}.}
}
\description{
Create a box dropdown item

Alias to \link{cardDropdownItem}
See \link{cardDropdownItem} for full details
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
