% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{dropdownItemList}
\alias{dropdownItemList}
\title{Create a box dropdown item list}
\usage{
dropdownItemList(..., icon = "wrench")
}
\arguments{
\item{...}{Slot for \link{dropdownItem}.}

\item{icon}{Dropdown menu icon.}
}
\description{
Can be used to add dropdown items to a cardtool.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
