% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{findSidebarItem}
\alias{findSidebarItem}
\title{Internally used by \link{bs4SidebarMenu} to find treeview items
and normal items.}
\usage{
findSidebarItem(items, regex)
}
\arguments{
\item{items}{List to search in.}

\item{regex}{Regex to apply.}
}
\description{
Internally used by \link{bs4SidebarMenu} to find treeview items
and normal items.
}
