% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{ionicon}
\alias{ionicon}
\title{BS4 ionicons}
\usage{
ionicon(name)
}
\arguments{
\item{name}{Name of icon. See \url{https://ionicons.com}.}
}
\description{
Create a ionicon.
}
\note{
Similar to the icon function from shiny.
}
\examples{
if(interactive()){
 library(shiny)
 library(bs4Dash)
 
 shiny::shinyApp(
   ui = bs4DashPage(
     navbar = bs4DashNavbar(),
     sidebar = bs4DashSidebar(),
     controlbar = bs4DashControlbar(),
     footer = bs4DashFooter(),
     title = "test",
     body = bs4DashBody(
      ionicon(name ="heart"),
      ionicon(name ="beer")
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
