% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{updateCard}
\alias{updateCard}
\alias{updateBox}
\title{Alias to \link{updatebs4Card}
See \link{updatebs4Card} for full details}
\usage{
updateCard(
  id,
  action = c("remove", "toggle", "toggleMaximize", "restore", "update"),
  options = NULL,
  session = shiny::getDefaultReactiveDomain()
)

updateBox(
  id,
  action = c("remove", "toggle", "toggleMaximize", "restore", "update"),
  options = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\description{
Alias to \link{updatebs4Card}
See \link{updatebs4Card} for full details

Alias to \link{updatebs4Card}
See \link{updatebs4Card} for full details
}
