% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R, R/cards.R, R/aliases.R
\name{updatebs4CardSidebar}
\alias{updatebs4CardSidebar}
\alias{bs4CardSidebar}
\alias{updateCardSidebar}
\title{Programmatically toggle a bs4Card sidebar}
\usage{
updatebs4CardSidebar(session, inputId)

bs4CardSidebar(
  ...,
  inputId = NULL,
  width = "25\%",
  background = "#333a40",
  startOpen = FALSE,
  icon = "cogs"
)

updateCardSidebar(session, inputId)
}
\arguments{
\item{session}{Shiny session object.}

\item{inputId}{Unique sidebar id. Useful if you want to use \link{updatebs4CardSidebar}.}

\item{...}{Sidebar content.}

\item{width}{Sidebar width in percentage. 25\% by default. A character value of any width CSS understands (e.g. "100px").}

\item{background}{Sidebar background color. Dark by default.}

\item{startOpen}{Whether the sidebar is open at start. FALSE by default.}

\item{icon}{Sidebar icon.}
}
\description{
To insert in the sidebar slot of \link{bs4Card}.
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 shinyApp(
  ui = bs4DashPage(
    sidebar_collapsed = FALSE,
    controlbar_collapsed = TRUE,
    enable_preloader = FALSE,
    navbar = bs4DashNavbar(skin = "dark"),
    body = bs4DashBody(
      bs4Card(
        title = "Closable Box with gradient", 
        closable = TRUE, 
        width = 12,
        height = "500px",
        solidHeader = FALSE, 
        collapsible = TRUE,
        actionButton("update", "Toggle card sidebar"),
        sidebar = bs4CardSidebar(
          inputId = "mycardsidebar",
          p("Sidebar Content")
        )
      )
    ),
    sidebar = bs4DashSidebar(),
    controlbar = bs4DashControlbar(),
    footer = bs4DashFooter()
  ),
  server = function(input, output, session) {
    observe(print(input$mycardsidebar))
    
    observeEvent(input$update, {
      updatebs4CardSidebar(session, inputId = "mycardsidebar")
    })
    
  }
 )
}
}
