% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{updateTabItems}
\alias{updateTabItems}
\title{Alias to \link{updatebs4TabItems}
See \link{updatebs4TabItems} for full details}
\usage{
updateTabItems(session, inputId, selected = NULL)
}
\description{
Alias to \link{updatebs4TabItems}
See \link{updatebs4TabItems} for full details
}
