% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userPostMedia}
\alias{userPostMedia}
\title{AdminLTE3 user post media}
\usage{
userPostMedia(src = NULL, height = NULL, width = NULL)
}
\arguments{
\item{height}{Media height in pixels.}

\item{width}{Media width in pixels.}

\item{image}{Image path or url ...}
}
\description{
\link{userPostMedia} creates a container to include an image in \link{userPost}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
